core/requirements.txt
Django==1.8
django-tastypie==0.12.1
django-tastypie-mongoengine==0.4.6
mongoengine==0.8.1
pymongo==2.8
python-dateutil==2.4.2
python-mimeparse==0.1.4
six==1.9.0
